10.07.2020

+++ database/migrations/2020_07_06_095800_create_blogs_table.php
@@ -15,9 +15,9 @@ class CreateBlogsTable extends Migration
     {
         Schema::create('blogs', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('cate_id')->unsigned();
+            $table->bigInteger('cate_id')->unsigned();
             $table->foreign('cate_id')->references('id')->on('blog_cates')->onDelete('cascade');
             $table->string('title');
             $table->longText('content');
